V3.01

+++ Script.sh
@@ -4,10 +4,11 @@ clear
 op=0
 op2=0
 name=0
-
+COL='\e[1;34m' #Color ${COL}
+NC='\033[0m' # No Color ${NC}
 ############ Estilos:
 #######
-# \033[1m Estilo normal \033[0m
+# \033[0m Estilo normal \033[0m
 # \033[1m Negrita \033[0m
 # \033[2m Dim \033[0m
 # \033[4m Subrayado \033[0m
@@ -17,22 +18,75 @@ name=0
 ######
 ############ Colores:
 ######
-# \e[0;30m Negro \e[0m
-# \e[1;30m Gris oscuro \e[0m
-# \e[0;34m Azul \e[0m
-# \e[1;34m Azlul resaltado \e[0m
-# \e[0;32m Verde \e[0m
-# \e[1;32m Verde resaltado \e[0m
-# \e[0;36m Cian \e[0m
-# \e[1;36m Cian resaltado \e[0m
-# \e[0;31m Rojo \e[0m
-# \e[1;31m Rojo resaltado \e[0m
-# \e[0;35m Púrpura \e[0m
-# \e[1;35m Púrpura resaltado \e[0m
-# \e[0;33m Cafe \e[0m
-# \e[1;33m Amarillo \e[0m
-# \e[0;37m Gris \e[0m
-# \e[1;37m Blanco \e[0m
+#Regular Colors
+#\033[0;30m  Black  \033[0m  # Black
+#\033[0;31m  Red    \033[0m  # Red
+#\033[0;32m  Green  \033[0m  # Green
+#\033[0;33m  Yellow \033[0m  # Yellow
+#\033[0;34m  Blue   \033[0m  # Blue
+#\033[0;35m  Purple \033[0m  # Purple
+#\033[0;36m  Purple \033[0m  # Cyan
+#\033[0;37m  White  \033[0m  # White
+
+#Bold
+#\033[1;30m  Black  \033[0m  # Black
+#\033[1;31m  Red    \033[0m  # Red
+#\033[1;32m  Green  \033[0m  # Green
+#\033[1;33m  Yellow \033[0m  # Yellow
+#\033[1;34m  Blue   \033[0m  # Blue
+#\033[1;35m  Purple \033[0m  # Purple
+#\033[1;36m  Purple \033[0m  # Cyan
+#\033[1;37m  White  \033[0m  # White
+
+#Underline
+#\033[4;30m  Black  \033[0m  # Black
+#\033[4;31m  Red    \033[0m  # Red
+#\033[4;32m  Green  \033[0m  # Green
+#\033[4;33m  Yellow \033[0m  # Yellow
+#\033[4;34m  Blue   \033[0m  # Blue
+#\033[4;35m  Purple \033[0m  # Purple
+#\033[4;36m  Purple \033[0m  # Cyan
+#\033[4;37m  White  \033[0m  # White
+
+#Background
+#\033[40m  Black  \033[0m    # Black
+#\033[41m  Red    \033[0m    # Red
+#\033[42m  Green  \033[0m    # Green
+#\033[43m  Yellow \033[0m    # Yellow
+#\033[44m  Blue   \033[0m    # Blue
+#\033[45m  Purple \033[0m    # Purple
+#\033[46m  Purple \033[0m    # Cyan
+#\033[47m  White  \033[0m    # White
+
+#High Intensity
+#\033[0;90m  Black  \033[0m  # Black
+#\033[0;91m  Red    \033[0m  # Red
+#\033[0;92m  Green  \033[0m  # Green
+#\033[0;93m  Yellow \033[0m  # Yellow
+#\033[0;94m  Blue   \033[0m  # Blue
+#\033[0;95m  Purple \033[0m  # Purple
+#\033[0;96m  Purple \033[0m  # Cyan
+#\033[0;97m  White  \033[0m  # White
+
+#Bold High Intensity
+#\033[1;90m  Black  \033[0m  # Black
+#\033[1;91m  Red    \033[0m  # Red
+#\033[1;92m  Green  \033[0m  # Green
+#\033[1;93m  Yellow \033[0m  # Yellow
+#\033[1;94m  Blue   \033[0m  # Blue
+#\033[1;95m  Purple \033[0m  # Purple
+#\033[1;96m  Purple \033[0m  # Cyan
+#\033[1;97m  White  \033[0m  # White
+
+#High Intensity backgrounds
+#\033[0;100m  Black  \033[0m  # Black
+#\033[0;101m  Red    \033[0m  # Red
+#\033[0;102m  Green  \033[0m  # Green
+#\033[0;103m  Yellow \033[0m  # Yellow
+#\033[0;104m  Blue   \033[0m  # Blue
+#\033[0;105m  Purple \033[0m  # Purple
+#\033[0;106m  Purple \033[0m  # Cyan
+#\033[0;107m  White  \033[0m  # White
 ######
 ############ To Do List:
 # ! Reparar la validacion de htop
@@ -85,11 +139,13 @@ while [ "$op" != 4 ]; do
             2) sleep 1s
                 last
                 sleep 1s
+                read -r -p "Presione enter para [ CONTINUAR ]..." op
+                [ "$op" = true ]
                 echo
             ;;
 
             3)
-                read -r -p "Presione cualquier tecla para [ CONTINUAR ]..." op
+                read -r -p"Presione enter para [ CONTINUAR ]..." op
                 [ "$op" = true ]
                 clear
             ;;
@@ -137,7 +193,7 @@ while [ "$op" != 4 ]; do
             ;;
 
             3) sleep 1s
-                sudo apt-get install htop
+                dpkg -V htop
                 sleep 3s
                 echo
                 htop -t
@@ -195,7 +251,7 @@ while [ "$op" != 4 ]; do
                         ;;
 
                         6)
-                            read -r -p "Presione cualquier tecla para [ CONTINUAR ]..." op
+                            read -r -p "Presione enter para [ CONTINUAR ]..." op
                             [ "$op" = true ]
                             clear
                         ;;
@@ -213,7 +269,7 @@ while [ "$op" != 4 ]; do
             ;;
 
             6) #Vuelve al inicio, puesto que al no haber una validación el while se vuelve nulo
-                read -r -p "Presione cualquier tecla para [ CONTINUAR ]..." op
+                read -r -p "Presione enter para [ CONTINUAR ]..." op
                 [ "$op" = true ]
                 clear
             ;;
@@ -237,16 +293,23 @@ while [ "$op" != 4 ]; do
         echo -e "\e[1;33m\033[1m|\e[0m 2. Iniciar aplicaciones              \e[1;33m\033[1m|\e[0m"
         echo -e "\e[1;33m\033[1m|\e[0m 3. Salir                             \e[1;33m\033[1m|\e[0m"
         echo -e "\e[1;33m\033[1m*--------------------------------------*\e[0m"
-        read -r -p "=> " op
+        read -r -p "=> " op2
         echo
         
-        case $op in 
+        case $op2 in 
             1) sleep 1s
                 echo "Para cerrar procesos sera necesario el numero PID el cual"
                 echo "es el, Identificad de procesos (Process Indentifiers). "
                 echo "Puedes abrir el administrador de tareas anterior y comprobar el PID"
                 read -r -p "PID => " pid
-                kill "$pid"
+
+                if [ "$pid" -le 893 ]; then
+                    echo -e "\e[1;35m\033[4mLa integridad del sistema se veria afectada\e[0m\033[0m"
+                elif [ "$pid" -le 1116 ]; then
+                    echo -e "\e[1;35m\033[4mLa integridad del sistema se veria afectada\e[0m\033[0m"
+                else
+                    kill "$pid"
+                fi
                 echo
             ;;
             
@@ -258,7 +321,7 @@ while [ "$op" != 4 ]; do
             ;;
 
             3)
-                read -r -p "Presione cualquier tecla para [ CONTINUAR ]..." op
+                read -r -p "Presione enter para [ CONTINUAR ]..." op
                 [ "$op" = true ]
                 clear
             ;;
